Minor bugs, PEP 514 support, execution args support

+++ src/manage/_core/pep514utils.py
@@ -1,6 +1,8 @@
 import time
 import winreg
 
+from .logging import LOGGER
+
 
 REG_TYPES = {
     str: winreg.REG_SZ,
@@ -8,7 +10,33 @@ REG_TYPES = {
 }
 
 
+class KeyNotFoundSentinel:
+    def __bool__(self):
+        return False
+
+    def __enter__(self):
+        return self
+
+    def __exit__(self, *args):
+        pass
+
+
+def _reg_open(root, subkey, writable=False, x86_only=None):
+    access = winreg.KEY_ALL_ACCESS if writable else winreg.KEY_READ
+    if x86_only is not None:
+        if x86_only:
+            access |= winreg.KEY_WOW64_32KEY
+        else:
+            access |= winreg.KEY_WOW64_64KEY
+    try:
+        return winreg.OpenKeyEx(root, subkey, access=access)
+    except FileNotFoundError:
+        return KeyNotFoundSentinel()
+
+
 def _iter_keys(key):
+    if not key:
+        return
     for i in range(0, 1024):
         try:
             yield winreg.EnumKey(key, i)
@@ -17,6 +45,8 @@ def _iter_keys(key):
 
 
 def _iter_values(key):
+    if not key:
+        return
     for i in range(0, 1024):
         try:
             yield winreg.EnumValue(key, i)
@@ -25,6 +55,8 @@ def _iter_values(key):
 
 
 def _delete_key(key, name):
+    if not key:
+        return
     for _ in range(5):
         try:
             winreg.DeleteKey(key, name)
@@ -36,6 +68,8 @@ def _delete_key(key, name):
 
 
 def _reg_rmtree(key, name):
+    if not key:
+        return
     try:
         subkey = winreg.OpenKey(key, name, access=winreg.KEY_ALL_ACCESS)
     except FileNotFoundError:
@@ -123,11 +157,7 @@ def update_registry(root_name, install, data):
 def cleanup_registry(root_name, keep):
     hive_name, _, root_name = root_name.partition("\\")
     hive = getattr(winreg, hive_name)
-    try:
-        root = winreg.OpenKey(hive, root_name, access=winreg.KEY_ALL_ACCESS)
-    except OSError:
-        return
-    with root:
+    with _reg_open(hive, root_name, writable=True) as root:
         for company_name in _iter_keys(root):
             any_left = False
             with winreg.OpenKey(root, company_name, access=winreg.KEY_ALL_ACCESS) as company:
@@ -138,3 +168,104 @@ def cleanup_registry(root_name, keep):
                         _reg_rmtree(company, tag_name)
             if not any_left:
                 _delete_key(root, company_name)
+
+
+def _read_str(key, value_name):
+    if not key:
+        return None
+    try:
+        v, vt = winreg.QueryValueEx(key, value_name)
+    except OSError:
+        return None
+    if vt == winreg.REG_SZ:
+        return v
+    if vt == winreg.REG_EXPAND_SZ:
+        import os
+        return os.path.expandvars(v)
+    return None
+
+
+def _read_one_unmanaged_install(company_name, tag_name, is_core, tag):
+    from pathlib import Path
+    from .verutils import Version
+
+    with _reg_open(tag, "InstallPath") as dirs:
+        prefix = _read_str(dirs, None)
+        exe = _read_str(dirs, "ExecutablePath")
+        exe_arg = _read_str(dirs, "ExecutableArguments")
+        exew = _read_str(dirs, "WindowedExecutablePath")
+        exew_arg = _read_str(dirs, "WindowedExecutableArguments")
+
+    display = _read_str(tag, "DisplayName")
+    ver = _read_str(tag, "Version")
+    
+    if not prefix or (not exe and not is_core):
+        raise ValueError("Registration is incomplete")
+    if is_core:
+        display = display or f"Python {tag_name}"
+        exe = exe or "python.exe"
+        exew = exew or "pythonw.exe"
+    if not ver:
+        ver = tag_name
+    while ver:
+        try:
+            Version(ver)
+            break
+        except Exception:
+            ver = ver[:-1]
+    else:
+        ver = "0"
+
+    prefix = Path(prefix)
+
+    i = {
+        "schema": 1,
+        "unmanaged": 1,
+        "id": f"__unmanaged-{company_name}-{tag_name}",
+        "sort-version": ver,
+        "company": company_name,
+        "tag": tag_name,
+        "run-for": [
+            {"tag": tag_name, "target": exe},
+        ],
+        "displayName": display or f"Unknown Python {company_name}\\{tag_name}",
+        "prefix": prefix,
+        "executable": prefix / exe,
+    }
+    if exe_arg:
+        i["run-for"][0]["args"] = exe_arg
+    if exew:
+        i["run-for"].append({"tag": tag_name, "target": exew, "windowed": 1})
+        if exew_arg:
+            i["run-for"][-1]["args"] = exew_arg
+    return i
+
+
+def _get_unmanaged_installs(root):
+    if not root:
+        return
+    for company_name in _iter_keys(root):
+        is_core = company_name.casefold() == "PythonCore".casefold()
+        with _reg_open(root, company_name) as company:
+            for tag_name in _iter_keys(company):
+                if _is_tag_managed(company, tag_name):
+                    continue
+                with _reg_open(company, tag_name) as tag:
+                    try:
+                        yield _read_one_unmanaged_install(company_name, tag_name, is_core, tag)
+                    except Exception:
+                        LOGGER.debug("Failed to read %s\\%s registration", company_name, tag_name)
+                        LOGGER.debug("ERROR", exc_info=True)
+
+
+def get_unmanaged_installs(sort_key=None):
+    installs = []
+    with _reg_open(winreg.HKEY_CURRENT_USER, "SOFTWARE\\Python") as root:
+        installs.extend(_get_unmanaged_installs(root))
+    with _reg_open(winreg.HKEY_LOCAL_MACHINE, "SOFTWARE\\Python", x86_only=False) as root:
+        installs.extend(_get_unmanaged_installs(root))
+    with _reg_open(winreg.HKEY_LOCAL_MACHINE, "SOFTWARE\\Python", x86_only=True) as root:
+        installs.extend(_get_unmanaged_installs(root))
+    if not sort_key:
+        return installs
+    return sorted(installs, key=sort_key)
